Don't clear terminal input queue in read
- Calling read(duration) doesn't clear the input queue when changing the
terminal settings. This is usually what we want, as this method is used
to read multi-byte character sequences. This fixes a problem we observed
on Scala Native

- Add a few tests for the CLong termios variant.

- Add a method to get special characters, used in testing.

- A bit of cleanup

+++ core/native/src/main/scala/terminus/NativeTerminal.scala
@@ -24,7 +24,7 @@ import scala.concurrent.duration.Duration
 import scala.scalanative.libc
 import scala.scalanative.meta.LinktimeInfo
 import scala.scalanative.posix
-import scala.scalanative.unsigned.UInt
+import scala.scalanative.unsigned.*
 
 import scalanative.unsafe.*
 
@@ -67,10 +67,10 @@ object NativeTerminal extends Terminal, WithEffect, TerminalKeyReader {
       val origAttrs = termios.getAttributes()
       val attrs = termios.getAttributes()
       try {
-        attrs.setSpecialCharacter(posix.termios.VMIN, 0)
+        attrs.setSpecialCharacter(posix.termios.VMIN, 0.toUByte)
         attrs.setSpecialCharacter(
           posix.termios.VTIME,
-          (duration.toMillis / 100).toByte
+          (duration.toMillis / 100).toUByte
         )
         termios.setAttributes(attrs)
 
@@ -131,48 +131,3 @@ object NativeTerminal extends Terminal, WithEffect, TerminalKeyReader {
     }
   }
 }
-
-// import scala.scalanative.posix.termios
-// import scala.scalanative.posix.unistd.STDIN_FILENO
-
-// object Termios {
-//   def getAttributes()(using Zone): Ptr[termios.termios] = {
-//     val attrs: Ptr[termios.termios] =
-//       alloc[TermiosStruct.clong_flags]()
-//     termios.tcgetattr(STDIN_FILENO, attrs)
-//     attrs
-//   }
-
-//   def setAttributes(ptr: Ptr[termios.termios]): Unit = {
-//     termios.tcsetattr(STDIN_FILENO, termios.TCSAFLUSH, ptr)
-//     // TODO: Error handling
-//     ()
-//   }
-
-//   /** Set the value at the given index of the termios c_cc struct member. Valid
-//     * indices are defined as constants such as `VMIN` and `VTIME` in
-//     * [[scala.scalanative.posix.termios]]
-//     */
-//   def setSpecialCharacter(
-//       attrs: Ptr[termios.termios],
-//       idx: CInt,
-//       value: CUnsignedChar
-//   ): Unit = {
-//     attrs._5(idx) = value
-//   }
-
-//   /** Places the terminal in raw mode.
-//     *
-//     * @see
-//     *   [[https://en.wikipedia.org/wiki/Terminal_mode Terminal Modes]]
-//     */
-//   def setRawMode(): Unit =
-//     Zone {
-//       println("Setting raw mode")
-//       val attrs = getAttributes()
-//       // attrs.removeLocalFlags(posix.termios.ECHO | posix.termios.ICANON)
-//       attrs._4 =
-//         attrs._4 & ~((posix.termios.ECHO | posix.termios.ICANON).toUInt)
-//       setAttributes(attrs)
-//     }
-// }
